разделил на модули функции
