+++ index.html
@@ -233,5 +233,8 @@
   </template>
   <script src="/js/functions.js"></script>
   <script src="/js/main.js"></script>
+  <script src="main.js" type="module"></script>
+  <script src="setup.js" type="module"></script>
+  <script src="data.js" type="module"></script>
 </body>
 </html>
